backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
alembic>=1.13
pydantic-settings>=2.2
psycopg[binary]>=3.1
PyJWT>=2.8
